online-outings/server/requirements.txt
asgiref==3.2.7
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.1
